+++ src/shared/types/Filter.type.ts
@@ -1,0 +1,4 @@
+export interface FilterOption {
+  id: string;
+  name: string;
+}
