+++ src/student/components/studentBenefits/benefitStudentHeader/BenefitStudentHeader.module.css
@@ -71,6 +71,7 @@
   font-weight: 500;
 }
 
+/* Responsive Design */
 @media (max-width: 768px) {
   .header {
     flex-direction: column;
